Improve Warlock/patron
The previous version was too close to deity. The new version relies on bonds to your patron which opens up lots of design space _and_ differentiates from the cleri's deity. Deity is impersonal and demands service, while patrons and personal and bargain. Your own personal Jesus, so to speak.

Thanks to Eric, Colin, and Chris for the ideas.